Fix compiler warning and conform to the new treatment of single floats.

+++ pypy/module/cppyy/src/reflexcwrapper.cxx
@@ -203,7 +203,7 @@ cppyy_methptrgetter_t cppyy_get_methptr_getter(cppyy_typehandle_t handle, int me
 /* handling of function argument buffer ----------------------------------- */
 void* cppyy_allocate_function_args(size_t nargs) {
     CPPYY_G__value* args = (CPPYY_G__value*)malloc(nargs*sizeof(CPPYY_G__value));
-    for (int i = 0; i < nargs; ++i)
+    for (size_t i = 0; i < nargs; ++i)
         args[i].type = 'l';
     return (void*)args;
 }
